# cmd/syncthing: Report real hashing performance, including weakhash