@scienmind 2017-01-19
What is the fundamental difference between the two kinds of measurements?
(seems like `sha256.SelectAlgo()` and `cpuBench()` called sequentially)