@calmh 2017-01-19
One is SHA256 only, the other is the full hash process which now is SHA256 and adler32 in sequence.