@AudriusButkevicius 2017-01-23
I suspect it's broken due to the newly added boolean to blocks indicating if we want weak hash or not.