@AudriusButkevicius 2017-01-23
@st-review merge