@st-review 2017-01-23
:ok_hand: Merged as 6d4fa27ea709e4c15b89353bd2fa50483ffac5e2. Thanks, @calmh!